+++ concurrency-checker-core/src/main/java/concurrencytest/checkpoint/Checkpoint2.java
@@ -4,13 +4,13 @@ import concurrencytest.annotations.InjectionPoint;
 
 public abstract class Checkpoint2 implements Checkpoint {
 
-    private final long id;
+    private final int id;
     private final InjectionPoint injectionPoint;
 
     private final String details, sourceFile;
     private final int lineNumber;
 
-    public Checkpoint2(long id, InjectionPoint injectionPoint, String details, String sourceFile, int lineNumber) {
+    public Checkpoint2(int id, InjectionPoint injectionPoint, String details, String sourceFile, int lineNumber) {
         this.id = id;
         this.injectionPoint = injectionPoint;
         this.details = details;
@@ -19,7 +19,7 @@ public abstract class Checkpoint2 implements Checkpoint {
     }
 
     @Override
-    public long checkpointId() {
+    public int checkpointId() {
         return id;
     }
